+++ packages/twenty-front/src/modules/activities/hooks/__tests__/useActivities.test.tsx
@@ -79,7 +79,7 @@ const mocks: MockedResponse[] = [
         }
       `,
       variables: {
-        filter: { activityTargetId: { eq: '123' } },
+        filter: { companyId: { eq: '123' } },
         limit: undefined,
         orderBy: undefined,
       },
@@ -180,7 +180,6 @@ describe('useActivities', () => {
           activitiesFilters: {},
           activitiesOrderByVariables: {},
           skip: false,
-          skipActivityTargets: false,
         }),
       { wrapper: Wrapper },
     );
@@ -188,7 +187,6 @@ describe('useActivities', () => {
     expect(result.current).toEqual({
       activities: [],
       loading: false,
-      noActivities: true,
     });
   });
 
@@ -201,12 +199,11 @@ describe('useActivities', () => {
 
         const activities = useActivities({
           targetableObjects: [
-            { targetObjectNameSingular: 'activityTarget', id: '123' },
+            { targetObjectNameSingular: 'company', id: '123' },
           ],
           activitiesFilters: {},
           activitiesOrderByVariables: {},
           skip: false,
-          skipActivityTargets: false,
         });
         return { activities, setCurrentWorkspaceMember };
       },
@@ -217,18 +214,9 @@ describe('useActivities', () => {
       result.current.setCurrentWorkspaceMember(mockWorkspaceMembers[0]);
     });
 
-    expect(result.current.activities.loading).toBe(true);
-
-    // Wait for activityTargets to complete fetching
-    await waitFor(() => !result.current.activities.loading);
-
-    expect(result.current.activities.loading).toBe(false);
-
-    // Wait for request to fetch activities to be made
-    await waitFor(() => result.current.activities.loading);
-
-    // Wait for activities to complete fetching
-    await waitFor(() => !result.current.activities.loading);
+    await waitFor(() => {
+      expect(result.current.activities.loading).toBe(false);
+    });
 
     const { activities } = result.current;
