+++ packages/twenty-front/src/modules/object-record/hooks/__tests__/useGenerateFindManyRecordsForMultipleMetadataItemsQuery.test.tsx
@@ -1,28 +1,17 @@
-import { ReactNode, useEffect, useState } from 'react';
+import { ReactNode } from 'react';
 import { expect } from '@storybook/test';
 import { renderHook } from '@testing-library/react';
-import { RecoilRoot, useSetRecoilState } from 'recoil';
+import { RecoilRoot } from 'recoil';
 
-import { objectMetadataItemsState } from '@/object-metadata/states/objectMetadataItemsState';
 import { getObjectMetadataItemsMock } from '@/object-metadata/utils/getObjectMetadataItemsMock';
 import { useGenerateCombinedFindManyRecordsQuery } from '@/object-record/multiple-objects/hooks/useGenerateCombinedFindManyRecordsQuery';
+import { JestObjectMetadataItemSetter } from '~/testing/jest/JestObjectMetadataItemSetter';
 
 const Wrapper = ({ children }: { children: ReactNode }) => (
   <RecoilRoot>
-    <ObjectMetadataItemSetter>{children}</ObjectMetadataItemSetter>
+    <JestObjectMetadataItemSetter>{children}</JestObjectMetadataItemSetter>
   </RecoilRoot>
 );
-
-const ObjectMetadataItemSetter = ({ children }: { children: ReactNode }) => {
-  const setObjectMetadataItems = useSetRecoilState(objectMetadataItemsState);
-  const [isLoaded, setIsLoaded] = useState(false);
-  useEffect(() => {
-    setObjectMetadataItems(getObjectMetadataItemsMock());
-    setIsLoaded(true);
-  }, [setObjectMetadataItems]);
-
-  return isLoaded ? <>{children}</> : null;
-};
 
 describe('useGenerateFindManyRecordsForMultipleMetadataItemsQuery', () => {
   it('should work as expected', async () => {
